+++ src/main/java/ru/coffeecoders/questbot/commands/admins/keybords/keyboards/creators/allQuestions/AllQuestionsKeyboardCreator.java
@@ -1,19 +1,23 @@
-package ru.coffeecoders.questbot.commands.admins.keybords.keyboards.creators;
+package ru.coffeecoders.questbot.commands.admins.keybords.keyboards.creators.allQuestions;
 
 import com.pengrad.telegrambot.model.request.InlineKeyboardButton;
 import com.pengrad.telegrambot.model.request.InlineKeyboardMarkup;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import ru.coffeecoders.questbot.commands.admins.keybords.keyboards.creators.Keyboard;
 
 @Component
-public class AllQuestionsKeyboardCreator extends InlineKeyboardMarkup implements Keyboard{
+public class AllQuestionsKeyboardCreator extends InlineKeyboardMarkup implements Keyboard {
     @Value("${keyboard.allQuestions.edit}")
     private static String edits;
 
-   @Value("${keyboard.allQuestions.toGroup}")
+   @Value("${keyboard.allQuestions.delete}")
     private static String deletes;
     private AllQuestionsKeyboardCreator allQuestionsKeyboardCreator;
 
+    @Value("${keyboard.allQuestions.show_news}")
+    static
+    String show_news;
 
     private AllQuestionsKeyboardCreator(AllQuestionsKeyboardCreator allQuestionsKeyboardCreator) {
         this.allQuestionsKeyboardCreator = allQuestionsKeyboardCreator;
@@ -36,6 +40,7 @@ public class AllQuestionsKeyboardCreator extends InlineKeyboardMarkup implements
         InlineKeyboardButton left = new InlineKeyboardButton("\t←");
         InlineKeyboardButton delete = new InlineKeyboardButton(deletes);
         InlineKeyboardButton edit = new InlineKeyboardButton(edits);
+        InlineKeyboardButton showNewList = new InlineKeyboardButton(show_news);
 
             return new InlineKeyboardButton[] {left, right,edit,delete};
     }
@@ -43,8 +48,9 @@ public class AllQuestionsKeyboardCreator extends InlineKeyboardMarkup implements
         InlineKeyboardButton [] firstRow = new InlineKeyboardButton [] {buttonArray[0], buttonArray[1]};
         InlineKeyboardButton [] secondRow = new InlineKeyboardButton [] {buttonArray[2]};
         InlineKeyboardButton [] thirdRow = new InlineKeyboardButton [] {buttonArray[3]};
+        InlineKeyboardButton [] fourthRow = new InlineKeyboardButton [] {buttonArray[4]};
 
-        return new InlineKeyboardButton [][] {firstRow, secondRow,thirdRow};
+        return new InlineKeyboardButton [][] {firstRow, secondRow,thirdRow, fourthRow};
     }
